Extract target validation tests into dedicated test file
Separate tests requiring external toolchains (cargo/bun/gradle) from
unit tests and self-hosting tests:

- Create tests/target_validation.rs with Rust/TS/Kotlin/Java self-hosted
  test execution (cargo test, bun test, gradle test)
- Move self_hosted_crate_content_check (string inspection) into
  self_hosting.rs where it belongs
- Remove self_hosting_compile.rs (split between target_validation.rs
  and self_hosting.rs)
- Remove guarantee_3_generated_*_tests_pass from self_host_guarantees.rs
  (moved to target_validation.rs)
- Update CLAUDE.md test table with 3-layer categorization:
  unit tests / self-hosting / target validation
